+++ src/program/GameLoop.cpp
@@ -552,12 +552,12 @@ void GameLoop::initProcessMessages()
     }
 
     /* Build and send the base savestate path/index */
-    if (context->config.sc.incremental_savestates) {
+    if (context->config.sc.savestate_settings & SharedConfig::SS_INCREMENTAL) {
         sendMessage(MSGN_BASE_SAVESTATE_INDEX);
         int index = 0;
         sendData(&index, sizeof(int));
 
-        if (!context->config.sc.savestates_in_ram) {
+        if (!(context->config.sc.savestate_settings & SharedConfig::SS_RAM)) {
             std::string basesavestatepath = context->config.savestatedir + '/';
             basesavestatepath += context->gamename;
             basesavestatepath += ".state0";
@@ -911,7 +911,7 @@ bool GameLoop::processEvent(uint8_t type, struct HotKey &hk)
             std::string savestatepath = context->config.savestatedir + '/';
             savestatepath += context->gamename;
             savestatepath += ".state" + std::to_string(statei);
-            if (! context->config.sc.savestates_in_ram) {
+            if (! (context->config.sc.savestate_settings & SharedConfig::SS_RAM)) {
                 sendMessage(MSGN_SAVESTATE_PATH);
                 sendString(savestatepath);
             }
@@ -1097,7 +1097,7 @@ bool GameLoop::processEvent(uint8_t type, struct HotKey &hk)
             }
 
             /* Send savestate path */
-            if (! context->config.sc.savestates_in_ram) {
+            if (! (context->config.sc.savestate_settings & SharedConfig::SS_RAM)) {
                 sendMessage(MSGN_SAVESTATE_PATH);
                 sendString(savestatepath);
             }